+++ src/pax/visible.odin
@@ -3,5 +3,5 @@ package pax
 Visible :: struct
 {
     sprite: int,
-    point:  [2]f32,
+    point:  [2]int,
 }
